0716 18:06 board paging complete

+++ src/main/java/com/project/ex01/BoardController.java
@@ -12,6 +12,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import service.CatBoardService;
 import vo.CatBoardVO;
+import vo.PageVO;
 
 @Controller
 public class BoardController {
@@ -19,14 +20,63 @@ public class BoardController {
 	CatBoardService service;
 	
 	@RequestMapping(value="catboard")
-	public ModelAndView catboard(ModelAndView mv) {
-		List<CatBoardVO> list = service.selectList();
+	public ModelAndView catboard(ModelAndView mv, PageVO<CatBoardVO> pvo) {
+		// ** paging 1 **
+		//1. paging 준비
+		// DAO의 pagelist를 처리하기 위해 필요한 값을 계산
+		// currentPage,startno,endno
+		int currentPage=1;
+		if(pvo.getCurrentPage()>1) {
+			currentPage=pvo.getCurrentPage();
+		}else {
+			pvo.setCurrentPage(currentPage);
+		}
+
+		int startRowno=(currentPage-1)*pvo.getPerPage()+1;
+		int endRowno=(startRowno+pvo.getPerPage())-1;
+		pvo.setStartno(startRowno);
+		pvo.setEndno(endRowno);
 		
-		mv.addObject("dnc",list);
+		//2) service
+		//DB에서 필요한 값들을 set
+		// 출력할 Row List, totalCount(totalRowCount)
+		pvo=service.pageList(pvo);
+		
+		//3) 결과처리
+		// totalCount 를 이용해서 totalPageNo 계산
+		// totalCount 가 70이면 totalPageNo는? (1page당 5개씩 출력)
+		//  73/5=14, 나머지가 있으면 +1
+		int totalPageNo = pvo.getTotalCount()/pvo.getPerPage();
+		if(pvo.getTotalCount()%pvo.getPerPage() >0)
+			totalPageNo+=1;
+		
+		//  ** paging 2 **
+		//sPageNo, ePageNo 계산
+		//필요한 값 : currentPage, perPageNo
+		//유형
+		//	1) 항상 현재 Page가 중앙에 위치할때
+		//		startPage : currentPage - perPageNo/2
+		//		endPage : currentPage - perPageNo/2
+		//
+		//	2) naver 카페 글, 11번가 상품 리스트 type
+		//		startPage : (((currentPage-1)/perPageNo)*perPageNo)+1
+		
+		int startPageNo=((currentPage-1)/pvo.getPerPageNO())*pvo.getPerPageNO()+1;
+		int endPageNo=startPageNo+pvo.getPerPageNO()-1;
+		
+		if(endPageNo>totalPageNo) endPageNo=totalPageNo;
+		
+		mv.addObject("startPage",startPageNo);
+		mv.addObject("endPage",endPageNo);
+		mv.addObject("perPageNO",pvo.getPerPageNO());
+		
+		mv.addObject("totalPageNo",totalPageNo);
+		mv.addObject("currentPage",currentPage);
+		
+		mv.addObject("list",pvo.getList());
 		mv.setViewName("cat/board/catboard");
-		
 		return mv;
-	} // boardlist
+	}// catboardpage
 	
 	@RequestMapping(value = "catboardinsertf")
 	public ModelAndView catboardinsertf(ModelAndView mv) {
@@ -89,7 +139,6 @@ public class BoardController {
 		}else {
 			mv.addObject("bcode",2);
 		}
-		
 		mv.setViewName("jsonView");
 		return mv;
 	}
@@ -97,15 +146,16 @@ public class BoardController {
 	@RequestMapping(value="catboarddelete")
 	public ModelAndView catboarddelete(HttpServletRequest request,ModelAndView mv, CatBoardVO bv) {
 		HttpSession session = request.getSession(false);
-		if(session!=null) {
+		if(session!=null && session.getAttribute("logID") != null) {
 			if(service.delete(bv)>0) {
-				mv.setViewName("redirect:catboard");
+				mv.addObject("bcode",0);
 			}else{
-				mv.setViewName("redirect:catboardview?seq="+bv.getSeq());
+				mv.addObject("bcode",1);
 			}
 		}else {
-			mv.setViewName("redirect:catmain?seq="+bv.getSeq());
+			mv.addObject("bcode",2);
 		}
+		mv.setViewName("jsonView");
 		return mv;
 	}
 } // class
